Rserve {Rserve}	R Documentation

Server providing R functionality to applications via TCP/IP or local unix sockets

Description

Starts Rserve in daemon mode (unix only). Any additional parameters not related to Rserve will be passed straight to the underlying R. For configuration, usage and command line parameters please consult the online documentation at http://www.rosuda.org/Rserve. Use R CMD Rserve --help for a brief help.

The Rserve function is provided for convenience only, it is recommended to start Rserve directly from the command line, not from R itself. Also note that the debug version of Rserve doesn't fork and thus will block until closed.

Usage


# R CMD Rserve [<parameters>]

Rserve(debug = FALSE, port = 6311, args = NULL)

Arguments

debug 	 determines whether regular Rserve or debug version of Rserve (Rserve.dbg) should be started.

port 	 port used by Rserve to listen for connections

args 	 further arguments passed to Rserve (as a string that will be passed to the system command thus use quotes where necessary).

Details

Rserve is not just a package, but an application. It is provided as a R package for convenience only. For details see http://www.rosuda.org/Rserve

Author(s)

Simon Urbanek

[Package Rserve version 0.6-0 Index]